practice/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
pycryptodome
argon2-cffi
pyjwt
pytest
locust
prometheus-fastapi-instrumentator
redis
python-json-logger
mmh3
cryptography
httpx
anyio
